Create utils/text.py
